requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
jenkinsapi==0.3.11
lxml==4.9.1
multi-key-dict==2.0.3
numpy==1.23.1
pandas==1.4.3
pbr==5.9.0
python-dateutil==2.8.2
python-jenkins==1.7.0
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.10
